+++ backend/canisters/storage_bucket/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Sync heap and stable memory sizes to StorageIndex ([#7192](https://github.com/open-chat-labs/open-chat/pull/7192))
 - Simplify handling of file expiration ([#7195](https://github.com/open-chat-labs/open-chat/pull/7195))
+- Use macro to create grouped timer job types ([#7224](https://github.com/open-chat-labs/open-chat/pull/7224))
 
 ## [[2.0.1560](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1560-storage_bucket)] - 2025-01-09